[CA-273985] Take clustering lock only if SM requires a cluster stack
Currently in xapi_pbd and xapi_sr we take a clustering lock (`with_clustering_lock`) regardless of the SR type.
However we only need to do this if the SM type requires a cluster stack. This commit adds a new function `with_clustering_lock_if_needed`, which calls `with_clustering_lock` if the SM type requires a cluster stack, otherwise it executes the passed function without any wrapper. The calls in xapi_pdb and xapi_sr are changed to use this new function.

This is tested by the fact that nesting two calls to `with_clustering_lock` will deadlock. The test creates one SM type with a required cluster stack and another without, and nests calls to `with_clustering_lock_if_needed` against each. If these both call `with_clustering_lock` (which is incorrect), we'll deadlock and the test will timeout and fail.

Signed-off-by: Callum McIntyre <callum.mcintyre@citrix.com>